requirements.txt
numpy==1.26.4
opencv-python-headless==4.9.0.80
scikit-learn==1.3.2
tensorflow-cpu==2.15.0
keras==2.15.0
fastapi
uvicorn
python-multipart
